Moved all remaining files out of opm/autodiff.
