+++ opm/simulators/flow/BlackoilModelEbos.hpp
@@ -34,7 +34,7 @@
 #include <opm/simulators/wells/BlackoilWellModel.hpp>
 #include <opm/simulators/aquifers/BlackoilAquiferModel.hpp>
 #include <opm/simulators/wells/WellConnectionAuxiliaryModule.hpp>
-#include <opm/autodiff/BlackoilDetails.hpp>
+#include <opm/simulators/flow/countGlobalCells.hpp>
 
 #include <opm/grid/UnstructuredGrid.h>
 #include <opm/core/simulator/SimulatorReport.hpp>
@@ -64,7 +64,6 @@
 #include <limits>
 #include <vector>
 #include <algorithm>
-//#include <fstream>
 
 
 BEGIN_PROPERTIES
