+++ opm/simulators/flow/countGlobalCells.hpp
