+++ tests/test_wellmodel.cpp
@@ -43,8 +43,6 @@
 #include <opm/grid/GridHelpers.hpp>
 #include <opm/simulators/flow/FlowMainEbos.hpp>
 #include <opm/simulators/flow/BlackoilModelEbos.hpp>
-#include <opm/autodiff/createGlobalCellArray.hpp>
-#include <opm/autodiff/GridInit.hpp>
 
 #include <ebos/eclproblem.hh>
 #include <ewoms/common/start.hh>
@@ -58,12 +56,13 @@
 #include <dune/common/parallel/mpihelper.hh>
 #endif
 
+
+
 using StandardWell = Opm::StandardWell<TTAG(EclFlowProblem)>;
 
 struct SetupTest {
 
     using Grid = UnstructuredGrid;
-    using GridInit = Opm::GridInit<Grid>;
 
     SetupTest ()
     {
@@ -81,12 +80,8 @@ struct SetupTest {
         const std::vector<double>& porv =
                             ecl_state->get3DProperties().getDoubleGridProperty("PORV").getData();
 
-        std::unique_ptr<GridInit> grid_init(new GridInit(*ecl_state, porv));
-        const Grid& grid = grid_init->grid();
-
-        // Create material law manager.
-        std::vector<int> compressed_to_cartesianIdx;
-        Opm::createGlobalCellArray(grid, compressed_to_cartesianIdx);
+        Opm::GridManager gm(ecl_state->getInputGrid(), porv);
+        const Grid& grid = *(gm.c_grid());
 
         current_timestep = 0;
